Move FormatUtil#formatTime to TimeUtil, and refactor
