+++ src/main/java/com/jagrosh/jmusicbot/utils/FormatUtil.java
@@ -25,19 +25,7 @@ import net.dv8tion.jda.core.entities.VoiceChannel;
  * @author John Grosh <john.a.grosh@gmail.com>
  */
 public class FormatUtil {
-    
-    public static String formatTime(long duration)
-    {
-        if(duration == Long.MAX_VALUE)
-            return "LIVE";
-        long seconds = Math.round(duration/1000.0);
-        long hours = seconds/(60*60);
-        seconds %= 60*60;
-        long minutes = seconds/60;
-        seconds %= 60;
-        return (hours>0 ? hours+":" : "") + (minutes<10 ? "0"+minutes : minutes) + ":" + (seconds<10 ? "0"+seconds : seconds);
-    }
-        
+
     public static String progressBar(double percent)
     {
         String str = "";
