+++ src/main/java/com/jagrosh/jmusicbot/utils/TimeUtil.java
@@ -18,4 +18,15 @@ package com.jagrosh.jmusicbot.utils;
 public class TimeUtil
 {
 
+    public static String formatTime(long duration)
+    {
+        if(duration == Long.MAX_VALUE)
+            return "LIVE";
+        long seconds = Math.round(duration/1000.0);
+        long hours = seconds/(60*60);
+        seconds %= 60*60;
+        long minutes = seconds/60;
+        seconds %= 60;
+        return (hours>0 ? hours+":" : "") + (minutes<10 ? "0"+minutes : minutes) + ":" + (seconds<10 ? "0"+seconds : seconds);
+    }
 }
